PDI-10657: Converted Rules Accumulator and Executor steps into a plugin. Drools 5.0.1.
