+++ build.properties
@@ -80,7 +80,7 @@ coverage.html.dir=${coverage.dir}/html
 
 includeantruntime=false
 
-plugin.list=kettle-gpload-plugin,kettle-hl7-plugin,kettle-palo-plugin,kettle-openerp-plugin,market,kettle5-log4j-plugin
+plugin.list=kettle-drools5-plugin,kettle-gpload-plugin,kettle-hl7-plugin,kettle-palo-plugin,kettle-openerp-plugin,market,kettle5-log4j-plugin
 module.list=core,engine,dbdialog,ui
 all-subprojects.list=${module.list},${plugin.list}
 assembly.stage.dir=${assembly.dir}/bin/stage
