+++ engine/src/kettle-steps.xml
@@ -156,8 +156,6 @@
  <step id="LoadFileInput"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.LoadFileInput</description> <classname>org.pentaho.di.trans.steps.loadfileinput.LoadFileInputMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Input</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.LoadFileInput</tooltip> <iconfile>ui/images/LFI.png</iconfile> <documentation_url/> <cases_url/> <forum_url/> </step>
  <step id="ChangeFileEncoding"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.ChangeFileEncoding</description> <classname>org.pentaho.di.trans.steps.changefileencoding.ChangeFileEncodingMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Utility</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.ChangeFileEncoding</tooltip> <iconfile>ui/images/CFE.png</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Change+file+encoding</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="FuzzyMatch"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.FuzzyMatch</description> <classname>org.pentaho.di.trans.steps.fuzzymatch.FuzzyMatchMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Lookup</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.FuzzyMatch</tooltip> <iconfile>ui/images/FZM.png</iconfile> <documentation_url/> <cases_url/> <forum_url/> </step>
- <step id="RuleExecutor"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.RuleExecutor</description> <classname>org.pentaho.di.trans.steps.rules.RulesExecutorMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Scripting</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.RuleExecutor</tooltip> <iconfile>ui/images/SCR_mod.png</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Rule+Executor</documentation_url> <cases_url/> <forum_url/> </step>
- <step id="RuleAccumulator"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.RuleAccumulator</description> <classname>org.pentaho.di.trans.steps.rules.RulesAccumulatorMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Scripting</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.RuleAccumulator</tooltip> <iconfile>ui/images/SCR_mod.png</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Rule+Accumulator</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="SSH"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.SSH</description> <classname>org.pentaho.di.trans.steps.ssh.SSHMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Utility</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.SSH</tooltip> <iconfile>ui/images/ssh.png</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Run+SSH+commands</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="RandomCCNumberGenerator"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.RandomCCNumberGenerator</description> <classname>org.pentaho.di.trans.steps.randomccnumber.RandomCCNumberGeneratorMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Input</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.RandomCCNumberGenerator</tooltip> <iconfile>ui/images/CCG.png</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Generate+random+credit+card+numbers</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="JsonInput"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.JsonInput</description> <classname>org.pentaho.di.trans.steps.jsoninput.JsonInputMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Input</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.JsonInput</tooltip> <iconfile>ui/images/JSI.png</iconfile> <documentation_url/> <cases_url/> <forum_url/> </step>
