+++ engine/src/org/pentaho/di/trans/step/messages/messages_en_US.properties
@@ -423,10 +423,6 @@ BaseStep.TypeLongDesc.ChangeFileEncoding=Change file encoding
 BaseStep.TypeTooltipDesc.ChangeFileEncoding=Change file encoding and create a new file
 BaseStep.TypeLongDesc.FuzzyMatch=Fuzzy match
 BaseStep.TypeTooltipDesc.FuzzyMatch=Finding approximate matches to a string using matching algorithms.\nRead a field from a main stream and output approximative value from lookup stream.
-BaseStep.TypeTooltipDesc.RuleExecutor=Execute a rule against each row or a set of rows
-BaseStep.TypeLongDesc.RuleExecutor=Rule Executor
-BaseStep.TypeTooltipDesc.RuleAccumulator=Execute a rule against a set of all incoming rows
-BaseStep.TypeLongDesc.RuleAccumulator=Rule Accumulator
 BaseStep.TypeLongDesc.SSH=Run SSH commands
 BaseStep.TypeTooltipDesc.SSH=Run SSH commands and returns result.
 BaseStep.TypeLongDesc.MetaInject=ETL Metadata Injection
