+++ engine/src/org/pentaho/di/trans/step/messages/messages_fr_FR.properties
@@ -118,7 +118,6 @@ BaseStep.TypeTooltipDesc.TypeExitGoogleAnalyticsStep=Extraction de donn\u00E9es 
 BaseStep.TypeTooltipDesc.TableExists=Cette \u00E9tape permet de v\u00E9rifier si une table\nexiste sur une connexion sp\u00E9cifi\u00E9e.
 BaseStep.Log.DistributionActivated=distribution activ\u00E9e
 BaseStep.TypeTooltipDesc.SingleThreader=Ex\u00E9cution d''une transformation sur un unique ex\u00E9tron.\nLa sous-transformation doit contenir une \u00E9tape d''injection par laquelle les donn\u00E9es arriveront depuis la transformation parent en bloc.
-BaseStep.TypeTooltipDesc.RuleExecutor=Ex\u00E9cution de r\u00E8gles sur un jeu de lignes
 BaseStep.TypeTooltipDesc.MongoDbInput=Extraction de toutes les entr\u00E9es depuis une base base Mongo.
 BaseStep.TypeLongDesc.AccessOutput=Alimentation base MS Access
 BaseStep.TypeTooltipDesc.SalesforceUpdate=Mise \u00E0 jour de donn\u00E9es dans un module Salesforce.
@@ -233,7 +232,6 @@ BaseStep.TypeLongDesc.XMLInputSax=Extraction depuis flux XML (SAX)
 BaseStepDialog.Connection.Label=Connexion 
 BaseStep.TypeTooltipDesc.OraBulkLoader=Cette \u00E9tape permet de charger en bloc\n des donn\u00E9es dans une base de donn\u00E9es Oracle
 BaseStep.ColumnName.LinesWritten=lignes_\u00E9crites
-BaseStep.TypeLongDesc.RuleExecutor=Ex\u00E9cution de r\u00E8gles
 BaseStep.TypeLongDesc.MongoDbInput=Extraction depuis une base Mongo
 BaseStep.TypeLongDesc.Dummy=Factice
 BaseStep.TypeTooltipDesc.Cubeinput=Lire donn\u00E9es depuis un fichier au format PDI.
