+++ engine/src/org/pentaho/di/trans/step/messages/messages_ja_JP.properties
@@ -417,8 +417,6 @@ BaseStep.TypeLongDesc.ChangeFileEncoding=\u30d5\u30a1\u30a4\u30eb\u6587\u5b57\u3
 BaseStep.TypeTooltipDesc.ChangeFileEncoding=Change file encoding\n\u30d5\u30a1\u30a4\u30eb\u306e\u30a8\u30f3\u30b3\u30fc\u30c7\u30a3\u30f3\u30b0\u3092\u5909\u66f4\u3057\u3001\u65b0\u305f\u306a\u30d5\u30a3\u30fc\u30eb\u30c9\u3092\u4f5c\u6210\u3057\u307e\u3059\u3002
 BaseStep.TypeLongDesc.FuzzyMatch=\u30d5\u30a1\u30b8\u30fc\u30de\u30c3\u30c1\u30f3\u30b0
 BaseStep.TypeTooltipDesc.FuzzyMatch=Fuzzy match\n\u30de\u30c3\u30c1\u30a2\u30eb\u30b4\u30ea\u30ba\u30e0\u3092\u4f7f\u7528\u3057\u3001\u8fd1\u4f3c\u4e00\u81f4\u3059\u308b\u6587\u5b57\u5217\u3092\u898b\u3064\u3051\u307e\u3059\u3002\n\u30e1\u30a4\u30f3\u30b9\u30c8\u30ea\u30fc\u30e0\u304b\u3089\u30d5\u30a3\u30fc\u30eb\u30c9\u3092\u8aad\u307f\u8fbc\u307f\u3001\u53c2\u7167\u30b9\u30c8\u30ea\u30fc\u30e0\u306e\u8fd1\u4f3c\u5024\u3092\u51fa\u529b\u3057\u307e\u3059\u3002
-BaseStep.TypeTooltipDesc.RuleExecutor=Rule Executor\n\u884c\u306b\u5bfe\u3057\u3066\u6307\u5b9a\u3057\u305f\u30eb\u30fc\u30eb\u3092\u5b9f\u884c\u3057\u307e\u3059\u3002
-BaseStep.TypeLongDesc.RuleExecutor=\u30eb\u30fc\u30eb\u5b9f\u884c
 BaseStep.TypeLongDesc.SSH=SSH\u30b3\u30de\u30f3\u30c9\u5b9f\u884c
 BaseStep.TypeTooltipDesc.SSH=Run SSH commands\nSSH\u30b3\u30de\u30f3\u30c9\u3092\u5b9f\u884c\u3057\u3066\u7d50\u679c\u3092\u53d6\u5f97\u3057\u307e\u3059\u3002
 BaseStep.TypeLongDesc.MetaInject=ETL\u30e1\u30bf\u30c7\u30fc\u30bf\u30a4\u30f3\u30b8\u30a7\u30af\u30b7\u30e7\u30f3
