+++ engine/src/org/pentaho/di/trans/step/messages/messages_zh_CN.properties
@@ -289,8 +289,6 @@ BaseStep.TypeLongDesc.ReservoirSampling=\u6570\u636e\u91c7\u6837
 BaseStep.TypeLongDesc.Rest=REST Client
 BaseStep.TypeLongDesc.RssInput=RSS \u8f93\u5165
 BaseStep.TypeLongDesc.RssOutput=RSS \u8f93\u51fa
-BaseStep.TypeLongDesc.RuleAccumulator=Rule Accumulator
-BaseStep.TypeLongDesc.RuleExecutor=Rule Executor
 BaseStep.TypeLongDesc.SASInput=SAS \u8f93\u5165
 BaseStep.TypeLongDesc.SQLFileOutput=SQL \u6587\u4ef6\u8f93\u51fa
 BaseStep.TypeLongDesc.SSH=\u8fd0\u884cSSH\u547d\u4ee4
