--- engine/src/org/pentaho/di/trans/steps/rules/messages/messages_en_US.properties
@@ -1,17 +1,0 @@
-RulesExecutor.StepConfigruationDialog.Title=Rules Executor
-
-RulesAccumulator.StepConfigruationDialog.Title=Rules Accumulator
-
-RulesData.Error.CompileDRL=Failed to compile DRL
-RulesData.InitializeColumns.InputRowMetaIsNull=No input rows found
-RulesDialog.StepName.Label=Step name
-RulesDialog.Tabs.RuleDefinition=Rules
-RulesDialog.Tabs.ColumnSelection=Results
-RulesDialog.RuleDefinition.FilePath=Rules file
-RulesDialog.RuleDefinition.RuleDefinition=Rules definition
-RulesDialog.ColumnSelection.ColumnName=Result column name
-RulesDialog.ColumnSelection.ColumnType=Result column type
-RulesMeta.Error.LoadFromXML=Unable to load step info from XML
-Dialog.Ok=OK
-Dialog.Cancel=Cancel
-
