+++ plugins/kettle-drools5-plugin/build.properties
@@ -1,0 +1,13 @@
+project.revision=TRUNK-SNAPSHOT
+ivy.artifact.group=pentaho-kettle
+ivy.artifact.id=kettle-drools5-plugin
+plugin-id=kettle-drools5-plugin
+impl.vendor=Pentaho Corporation
+impl.title=Drools5 Plugin
+impl.productID=${ivy.artifact.id}
+dev-lib.dir=dev-lib
+
+dependency.kettle.revision=${project.revision}
+dependency.pentaho-xul.revision=TRUNK-SNAPSHOT
+
+
