+++ plugins/kettle-drools5-plugin/build.xml
@@ -1,0 +1,94 @@
+<!--===========================================================================
+  This is the build file for the Pentaho Data Integration (Kettle) core module.
+  
+  This build file will use the subfloor.xml file as the default build
+  process and should only override the tasks that need to differ from
+  the common build file.
+  
+  See common_build.xml for more details
+============================================================================-->
+<project name="kettle-drools5-plugin" basedir="." default="default" xmlns:ivy="antlib:org.apache.ivy.ant">
+
+  <description>
+    This build file is used to create the DROOLS Kettle plugin
+    and works with the subfloor.xml file.
+  </description>
+
+  <!-- The continuous target is used by CI ... this is the list of -->
+  <!-- tasks that the CI machine will run.                         -->
+  <!-- DO NOT change the CI machine's ant task .. change this list -->
+  <target name="continuous" 
+          depends="clean-all,resolve,publish" />
+
+  <!-- Import the common_build.xml file which contains all the default tasks -->
+  <import file="build-res/subfloor.xml"/>
+
+  <!--
+    AS STATED ABOVE, THE ONLY TASKS THAT SHOULD EXIST IN THIS BUILD FILE ARE
+    THE TASKS THAT NEED TO DIFFER FROM THE DEFAULT IMPLEMENTATION OF THE TASKS
+    FOUND IN common_build.xml.
+  --> 
+
+  <target name="default" depends="clean,init,resolve-dev,dist"/>
+      
+  <target name="create-dot-classpath" depends="init,resolve,subfloor.create-dot-classpath"/>
+    
+  <target name="stage" depends="dist" description="generate all the kettle plugin jars">
+      
+  	  <echo>Staging the Kettle plugin ${plugin} ...</echo>
+  	
+      <!-- copy plugin jar -->
+      <mkdir dir="bin/stage/${plugin-id}"/>
+      <copy todir="bin/stage/${plugin-id}"  file="dist/${plugin-id}-${project.revision}.jar"/>
+      
+  	  <!-- external libs (jdbc etc) -->
+      <mkdir dir="bin/stage/${plugin-id}/lib"/>
+  	  <if>
+        <available file="lib" type="dir" />
+        <then>
+        	<echo>Copying files from lib....</echo>
+  	        <copy todir="bin/stage/${plugin-id}/lib">
+  	           <fileset dir="lib" includes="**/*"/>
+      	    </copy>
+  	  	</then>
+  	  </if>
+  	
+      <!-- any other res files -->
+      <if>
+          <available file="res" />
+          <then>
+             <copy todir="bin/stage/${plugin-id}" >
+                        <fileset dir="res" includes="**/*" excludes="**/.vpn"/>
+             </copy>
+            </then>
+          </if>
+                
+            <!-- Update the version.xml with the current version of this plugin -->
+            <if>
+              <available file="bin/stage/${plugin-id}/version.xml" />
+              <then>
+                <replace file="bin/stage/${plugin-id}/version.xml" token="@TRUNK@" value="${project.revision}"/>
+              </then>
+            </if>
+    </target>
+    
+   <!--=======================================================================
+            resolve-swtlibs
+            
+            Resolved the swtlibs configuration from ivy.xml
+       =====================================================================-->
+      <target name="resolve-dev" depends="install-ivy, resolve">
+          <ivy:resolve file="${ivyfile}" conf="dev" />
+	      <ivy:retrieve conf="dev" pattern="${dev-lib.dir}/[module]-[revision](-[classifier]).[ext]" />
+		  <touch file="${dev-lib.dir}/.kettle-ignore"/>
+	  </target>
+	
+  <target name="package" depends="stage" description="generate all the kettle plugin jars">
+     <echo>Creating the Kettle plugin zip for @{plugin} ...</echo>
+     <zip destfile="dist/${plugin-id}.zip" 
+          basedir="bin/stage/${plugin-id}"
+          includes="**/*"
+          excludes="**/dev-lib/*" />
+  </target>
+
+</project>
