+++ plugins/kettle-drools5-plugin/ivy.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ivy-module version="2.0" xmlns:m="http://ant.apache.org/ivy/maven">
+  <info organisation="pentaho-kettle" module="${ivy.artifact.id}" revision="${dependency.kettle.revision}" status="integration" publication="20140207120000"/>
+
+  <configurations>
+    <conf name="default"/>
+    <conf name="dev"/>
+    <conf name="test" visibility="private"/>
+    <conf name="source"/>
+    <conf name="zip"/>
+  </configurations>
+
+  <dependencies defaultconf="default->default">
+    <dependency org="pentaho-kettle" name="kettle-core" rev="${dependency.kettle.revision}" conf="dev->default" changing="true"/>
+    <dependency org="pentaho-kettle" name="kettle-engine" rev="${dependency.kettle.revision}" conf="dev->default" changing="true"/>
+    <dependency org="pentaho-kettle" name="kettle-ui-swt" rev="${dependency.kettle.revision}" conf="dev->default" changing="true"/>  
+
+    <dependency org="janino" name="janino" rev="2.5.16" transitive="false"/>
+
+    <dependency org="org.drools" name="drools-api" rev="5.0.1"/>
+    <dependency org="org.drools" name="drools-core" rev="5.0.1"/>
+    <dependency org="org.drools" name="drools-compiler" rev="5.0.1"/>
+    <dependency org="org.mvel" name="mvel2" rev="2.0.10"/>
+    <dependency org="org.eclipse.jdt" name="core" rev="3.4.2.v_883_R34x"/>    
+    
+    <dependency org="junit" name="junit" rev="4.7" conf="test->default"/>     
+  </dependencies>
+</ivy-module>
