+++ plugins/kettle-drools5-plugin/ivysettings.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ivysettings>
+  <properties environment="env" />
+  <property name="ivy.local.default.root" value="${ivy.default.ivy.user.dir}/local" override="true" />
+  <property name="ivy.local.default.artifact.pattern" value="[organisation]/[module]/[revision]/[type]s/[artifact]-[revision].[ext]"
+    override="false" />
+
+  <settings defaultResolver="pentaho-chained-resolver" />
+  <include url="${ivy.default.settings.dir}/ivysettings-public.xml" />
+  <include url="${ivy.default.settings.dir}/ivysettings-local.xml" />
+  <resolvers>
+    <chain name="pentaho-chained-resolver">
+      <resolver ref="local" />
+      <dual name="pentaho">
+        <url name="pentaho-ivy">
+          <ivy pattern="http://repo.pentaho.org/artifactory/repo/[organisation]/[module]/[revision]/[module]-[revision].ivy.xml" />
+        </url>
+        <ibiblio name="pentaho-mvn" m2compatible="true" root="http://repo.pentaho.org/artifactory/repo" />
+      </dual>
+      <resolver ref="public" />
+      <ibiblio name="java-net-maven2" root="http://download.java.net/maven/2/" m2compatible="true" />
+    </chain>
+  </resolvers>
+  <caches lockStrategy="artifact-lock" resolutionCacheDir="${ivy.default.ivy.user.dir}/resol-cache${env.EXECUTOR_NUMBER}" />
+</ivysettings>
