+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/Rules.java
@@ -85,6 +85,20 @@ public class Rules {
       this.external = external;
     }
 
+    public Column( String name, String type, Object payload  ) {
+      this();
+      this.name=name;
+      this.type=type;
+      this.payload=payload;
+    }
+
+    public Column( Boolean external, String name, String type, Object payload  ) {
+      this(external);
+      this.name=name;
+      this.type=type;
+      this.payload=payload;
+    }
+
     public String getName() {
       return name;
     }
