+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/RulesAccumulator.java
