+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/RulesAccumulatorData.java
@@ -23,52 +23,44 @@
 package org.pentaho.di.trans.steps.rules;
 
 import java.io.StringReader;
+import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
+import java.util.Hashtable;
 import java.util.List;
 import java.util.Map;
 
 import org.drools.KnowledgeBase;
 import org.drools.KnowledgeBaseFactory;
+import org.drools.ObjectFilter;
 import org.drools.builder.KnowledgeBuilder;
 import org.drools.builder.KnowledgeBuilderFactory;
 import org.drools.builder.ResourceType;
 import org.drools.definition.KnowledgePackage;
 import org.drools.io.Resource;
 import org.drools.io.ResourceFactory;
-import org.drools.runtime.ObjectFilter;
 import org.drools.runtime.StatefulKnowledgeSession;
 import org.pentaho.di.core.row.RowMetaInterface;
-import org.pentaho.di.core.row.ValueMetaInterface;
 import org.pentaho.di.i18n.BaseMessages;
 import org.pentaho.di.trans.step.BaseStepData;
 import org.pentaho.di.trans.step.StepDataInterface;
-import org.pentaho.di.trans.steps.rules.Rules.Column;
+import org.pentaho.di.trans.steps.rules.Rules.Row;
 
-/**
- * This Transformation Step allows a user to execute a rule set against an individual rule or a collection of rules.
- *
- * Additional columns can be added to the output from the rules and these (of course) can be used for routing if
- * desired.
- *
- * @author cboyden
- *
- */
-
-public class RulesExecutorData extends BaseStepData implements StepDataInterface {
-  private static Class<?> PKG = RulesExecutor.class; // for i18n purposes
+public class RulesAccumulatorData extends BaseStepData implements StepDataInterface {
+  private static Class<?> PKG = RulesAccumulator.class; // for i18n purposes
 
   private RowMetaInterface outputRowMeta;
+  private RowMetaInterface inputRowMeta;
 
   private KnowledgeBuilder kbuilder;
 
   private KnowledgeBase kbase;
 
-  private Column[] columnList;
-
-  private Map<String, Column> resultMap = new HashMap<String, Column>();
+  private List<Object[]> results;
 
   private String ruleString;
+
+  private List<Row> rowList = new ArrayList<Row>();
+  private List<Row> resultList = new ArrayList<Row>();
 
   public String getRuleString() {
     return ruleString;
@@ -97,6 +89,11 @@ public class RulesExecutorData extends BaseStepData implements StepDataInterface
   }
 
   public void initializeRules() {
+
+    // To ensure the plugin classloader use for dependency resolution
+    ClassLoader loader = getClass().getClassLoader();
+    Thread.currentThread().setContextClassLoader(loader);
+
     Resource ruleSet = null;
     if ( ruleString != null ) {
       ruleSet = ResourceFactory.newReaderResource( new StringReader( ruleString ) );
@@ -118,82 +115,66 @@ public class RulesExecutorData extends BaseStepData implements StepDataInterface
     kbase.addKnowledgePackages( pkgs );
   }
 
-  public void initializeColumns( RowMetaInterface inputRowMeta ) {
-    if ( inputRowMeta == null ) {
+  public void initializeInput( RowMetaInterface _inputRowMeta ) {
+    if ( _inputRowMeta == null ) {
       BaseMessages.getString( PKG, "RulesData.InitializeColumns.InputRowMetaIsNull" );
       return;
     }
 
-    // Create objects for insertion into the rules engine
-    List<ValueMetaInterface> columns = inputRowMeta.getValueMetaList();
-
-    // This array must 1-1 match the row[] feteched by getRow()
-    columnList = new Column[columns.size()];
-
-    for ( int i = 0; i < columns.size(); i++ ) {
-      ValueMetaInterface column = columns.get( i );
-
-      Column c = new Column( true );
-      c.setName( column.getName() );
-      c.setType( column.getTypeDesc() );
-      c.setPayload( null );
-
-      columnList[i] = c;
-    }
+    this.inputRowMeta = _inputRowMeta;
   }
 
-  public void loadRow( Object[] r ) {
-    for ( int i = 0; i < columnList.length; i++ ) {
-      columnList[i].setPayload( r[i] );
-    }
-    resultMap.clear();
-  }
-
-  public void execute() {
-    StatefulKnowledgeSession session = initNewKnowledgeSession();
-
-    Collection<Object> oList = fetchColumns( session );
-    for ( Object o : oList ) {
-      resultMap.put( ( (Column) o ).getName(), (Column) o );
+  public void loadRow( Object[] r ) throws Exception {
+    // Store rows for processing
+    Map<String, Object> columns = new Hashtable<String, Object>();
+    for ( String field : inputRowMeta.getFieldNames() ) {
+      columns.put( field, r[inputRowMeta.indexOfValue( field )] );
     }
 
-    session.dispose();
+    rowList.add( new Row( columns, true ) );
   }
 
-  protected StatefulKnowledgeSession initNewKnowledgeSession() {
-    StatefulKnowledgeSession session = kbase.newStatefulKnowledgeSession();
-    for ( int i = 0; i < columnList.length; i++ ) {
-      session.insert( columnList[i] );
-    }
-
-    session.fireAllRules();
-    return session;
+  public List<Row> getResultRows() {
+    return resultList;
   }
 
-  protected Collection<Object> fetchColumns( StatefulKnowledgeSession session ) {
-    Collection<Object> oList = session.getObjects( new ObjectFilter() {
-      @Override
-      public boolean accept( Object o ) {
-        if ( o instanceof Column && !( (Column) o ).isExternalSource() ) {
-          return true;
-        }
-        return false;
+  public void execute() throws Exception {
+    if ( kbase != null ) {
+      StatefulKnowledgeSession session = kbase.newStatefulKnowledgeSession();
+
+      for ( Row row : rowList ) {
+        session.insert( row );
       }
-    } );
-    return oList;
+
+      session.fireAllRules();
+
+      Collection<Object> oList = session.getObjects( new ObjectFilter() {
+        @Override
+        public boolean accept( Object o ) {
+          if ( o instanceof Row && !( (Row) o ).isExternalSource() ) {
+            return true;
+          }
+          return false;
+        }
+      } );
+
+      for ( Object o : oList ) {
+        resultList.add( (Row) o );
+      }
+
+      session.dispose();
+    }
   }
 
   /**
+   * Get the list of rows generated by the Rules execution
    *
-   * @param columnName
-   *          Column.payload associated with the result, or null if not found
-   * @return
+   * @return List of rows generated
    */
-  public Object fetchResult( String columnName ) {
-    return resultMap.get( columnName );
+  public List<Object[]> fetchResults() {
+    return results;
   }
 
   public void shutdown() {
   }
-
 }
