+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/messages/messages_en_US.properties
@@ -1,0 +1,19 @@
+RulesExecutor.StepConfigruationDialog.Title=Rule Executor
+RulesExecutor.StepConfigruationDialog.TooltipDesc=Execute a rule against each row
+
+RulesAccumulator.StepConfigruationDialog.Title=Rules Accumulator
+RulesAccumulator.StepConfigruationDialog.TooltipDesc=Execute a rule against a set of all incoming rows
+
+RulesData.Error.CompileDRL=Failed to compile DRL
+RulesData.InitializeColumns.InputRowMetaIsNull=No input rows found
+RulesDialog.StepName.Label=Step name
+RulesDialog.Tabs.RuleDefinition=Rules
+RulesDialog.Tabs.ColumnSelection=Results
+RulesDialog.RuleDefinition.FilePath=Rules file
+RulesDialog.RuleDefinition.RuleDefinition=Rules definition
+RulesDialog.ColumnSelection.ColumnName=Result column name
+RulesDialog.ColumnSelection.ColumnType=Result column type
+RulesMeta.Error.LoadFromXML=Unable to load step info from XML
+Dialog.Ok=OK
+Dialog.Cancel=Cancel
+
