+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/messages/messages_fr_FR.properties
@@ -1,0 +1,7 @@
+#File generated by Pentaho Translator for package 'org.pentaho.di.trans.steps.rules' in locale 'fr_FR'
+RulesData.InitializeColumns.InputRowMetaIsNull=Aucun flux d''entr\u00E9e n''a \u00E9t\u00E9 d\u00E9tect\u00E9\!
+RulesMeta.Error.LoadFromXML=Erreur lors du chargement depuis le fichier XML, des informations de l''\u00E9tape
+RulesData.Error.CompileDRL=Echec lors de la compilation DRL
+
+RulesExecutor.StepConfigruationDialog.Title=Ex\u00E9cution de r\u00E8gles
+RulesExecutor.StepConfigruationDialog.TooltipDesc=Ex\u00E9cution de r\u00E8gles sur un jeu de lignes
