+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/messages/messages_it_IT.properties
