+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/messages/messages_ja_JP.properties
@@ -1,0 +1,14 @@
+RulesData.InitializeColumns.InputRowMetaIsNull=\u5165\u529b\u884c\u304c\u898b\u3064\u304b\u308a\u307e\u305b\u3093\u3067\u3057\u305f\u3002
+StepConfigruationDialog.Title=\u30eb\u30fc\u30eb\u5b9f\u884c
+RulesDialog.StepName.Label=\u30b9\u30c6\u30c3\u30d7\u540d
+RulesDialog.Tabs.RuleDefinition=\u30eb\u30fc\u30eb
+RulesDialog.Tabs.ColumnSelection=\u7d50\u679c
+RulesDialog.RuleDefinition.FilePath=\u30eb\u30fc\u30eb\u30d5\u30a1\u30a4\u30eb
+RulesDialog.RuleDefinition.RuleDefinition=\u30eb\u30fc\u30eb\u5b9a\u7fa9
+RulesDialog.ColumnSelection.ColumnName=\u7d50\u679c\u5217\u540d
+RulesDialog.ColumnSelection.ColumnType=\u7d50\u679c\u5217\u30bf\u30a4\u30d7
+Dialog.Ok=\u6210\u529f
+Dialog.Cancel=\u30ad\u30e3\u30f3\u30bb\u30eb
+
+RulesExecutor.StepConfigruationDialog.Title=\u30eb\u30fc\u30eb\u5b9f\u884c
+RulesExecutor.StepConfigruationDialog.TooltipDesc=Rule Executor\n\u884c\u306b\u5bfe\u3057\u3066\u6307\u5b9a\u3057\u305f\u30eb\u30fc\u30eb\u3092\u5b9f\u884c\u3057\u307e\u3059\u3002
