+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/trans/steps/rules/messages/messages_ko_KR.properties
@@ -1,0 +1,18 @@
+#Generated by ResourceBundle Editor (http://eclipse-rbe.sourceforge.net)
+#File generated by Pentaho Translator for package 'org.pentaho.di.trans.steps.rules' in locale 'ko_KR'
+#
+#
+#Mon Jul 12 10:06:26 KST 2010
+
+Dialog.Cancel = \uCDE8\uC18C
+Dialog.Ok     = \uC801\uC6A9
+
+RulesData.InitializeColumns.InputRowMetaIsNull = \uC785\uB825 \uB85C\uC6B0\uAC00 \uC5C6\uC2B5\uB2C8\uB2E4
+
+RulesDialog.ColumnSelection.ColumnName    = \uACB0\uACFC \uCEEC\uB7FC \uC774\uB984
+RulesDialog.ColumnSelection.ColumnType    = \uACB0\uACFC \uCEEC\uB7FC \uD615\uC2DD
+RulesDialog.RuleDefinition.FilePath       = \uB8F0 \uD30C\uC77C
+RulesDialog.RuleDefinition.RuleDefinition = \uB8F0 \uC815\uC758
+RulesDialog.StepName.Label                = Step \uC774\uB984
+RulesDialog.Tabs.ColumnSelection          = \uACB0\uACFC
+RulesDialog.Tabs.RuleDefinition           = \uB8F0
