+++ plugins/kettle-drools5-plugin/src/org/pentaho/di/ui/trans/steps/rules/RulesExecutorMetaMapper.java
